refactor: 트랜잭션 매니저 분리
